Found in update statistics process and reported by YourSqlDba when it occurs.
Workaround for a rare case where sys.dm_physical_index_stats returns an object_id for which object_schema_name(object_d, db_id) function resolves to a null value.  This then raise a bug when inserting this data to temporary table because insert with a null value is rejected.  

Solution : Just discard rows for which this condition occurs.