Version 6.1
This version remove a bug and another annoyance related to a backup myth.

It removes bug truncation error when deleting very long backup names in Maint.DeleteOldBackups procedure. One variable had to by made larger.

It removes a mechanism to exclude simultaneous log backups with other database maintenance operations on the same database. We discovered that this restriction was lifted since SQL2005. To see a reliable comment about this go to this article. http://www.sqlskills.com/blogs/paul/search-engine-qa-16-concurrent-log-and-full-backups/. Because of this mechanism, YourSqlDba was sometime giving errors because of too long log backups. So it wouldn't be a problem anymore.

Version 6.1.1

1) Correction of string truncation error when deleting very long backup names in Maint.DeleteOldBackups procedure.

2) Removal of the mechanism that exclude simulatenous Log backups with other database maintenance operations on the same database. We discovered that this requirement was lifted since SQL2005. To see a reliable comment about this go to this article. http://www.sqlskills.com/blogs/paul/search-engine-qa-16-concurrent-log-and-full-backups/.
Because of this mechanism, YourSqlDba was sometime giving errors because excessive wait on long log backups, so this wont' happen anymore.

3) YourSqlDba relies on logreusewait status in sys.databases to do log shrink attempts. If this status was <> 0 a 10 seconds delay was performed. It appears that, in SQL2012, this status is 2 most of the time (log backup), which makes this delay appears at every attempts to shrink a log's database. This status doesn't cause any error at shrink attempt, so we include it in the list of status that doesn't prevent shrink log attempts.

Version 6.2

This version proposes a more transparent and user-friendly integration of YourSqlDba account with mirroring. 

Upon deployment this version checks that YourSqlDba account has access on each Mirror. If a single one of them fails, it tries to repair automatically this link This should always works provided that your actual connection has sysadmin access to the mirror server through the matching linked server name. The same is true also for maintenance tasks.

However if you use a single mirror server for multiple source servers, we encourage to set yourself this password on each source servers with the new stored procedure Mirroring.SetYourSqlDbaAccountForMirroring which takes @SetYourSqlDbaPasswordPwd parameter to specify a password of your choice. By doing so linked server access through each source servers will work, and this would avoid automatic constant repair of the link at each maintenance.

This version modifies local YourSqlDba account password only when SetYourSqlDbaPassword is called with @SetYourSqlDbaPasswordPwd completed. When not specified the modification is done only at the mirror with a randomly generated password, which is enough when using single source server for a given mirror server.

This should solve any deployment issues since the introduction of this account.