Update YourSQLDba_InstallOrUpdateScript.sql
This version implement a new parameter for mirroring called @migrationTestMode.
It's goal is just to set mirroring as a simple mean to backup/restore to another server, e.g. in a migration to a newer SQL server as set of database, and put them online. While they exist and are online, YourSqlDba do not attempt another restore of any kind.  To restore normal YourSqlDba mirroring, just remove the @migrationTestMode=1 parameter and suppress test databases on target server.

In another field, Install.InitialSetupOfYourSqlDba was improved to send a email on successful install.  If this email reach its destination, it proves that email configuration is ok.