Version 6.7.0.1 Introduces a better way to display YourSqlDba job history and some minor fixes
A new function helps to display YourSqlDba logs. It's name is Maint.History view.  YourSqlDba job report provides a sample about how to use it, and ReadMe provides a link to it description.  This version also fixes a minor problem that made jobs history details to be lost at update.